moved Details component into alert-details.tsx file
